requirements.txt
trafilatura~=2.0.0
pydantic~=2.11.7
python-dotenv~=1.1.1
click~=8.1.7
google-genai~=1.36.0